Fix Python Emitter
