+++ src/compiler/emitters/python/src/commonMain/kotlin/community/flock/wirespec/emitters/python/PythonIdentifierEmitter.kt
@@ -1,0 +1,25 @@
+package community.flock.wirespec.emitters.python
+
+import community.flock.wirespec.compiler.core.emit.Emitter.Companion.firstToUpper
+import community.flock.wirespec.compiler.core.emit.IdentifierEmitter
+import community.flock.wirespec.compiler.core.parse.DefinitionIdentifier
+import community.flock.wirespec.compiler.core.parse.FieldIdentifier
+import community.flock.wirespec.compiler.core.parse.Identifier
+import community.flock.wirespec.emitters.python.PythonEmitter.Companion.reservedKeywords
+
+interface PythonIdentifierEmitter: IdentifierEmitter {
+
+    override fun emit(identifier: Identifier) = when (identifier) {
+        is DefinitionIdentifier -> identifier.sanitize()
+        is FieldIdentifier -> identifier.sanitize().sanitizeKeywords()
+    }
+
+    override fun Identifier.sanitize() = value
+        .split(".", " ")
+        .mapIndexed { index, s -> if (index > 0) s.firstToUpper() else s }
+        .joinToString("")
+        .filter { it.isLetterOrDigit() || it == '_' }
+        .let { if (it.firstOrNull()?.isDigit() == true) "_$it" else it }
+
+    fun String.sanitizeKeywords() = if (this in reservedKeywords) "_$this" else this
+}
